+++ libs/node-std/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @palmares/node-std
 
+## 0.1.2
+
+### Patch Changes
+
+- asdasdsa
+- Updated dependencies
+  - @palmares/std@0.1.2
+
+## 0.1.1
+
+### Patch Changes
+
+- Testing if changesets and if custom script works
+- Updated dependencies
+  - @palmares/std@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes