+++ libs/redis-emitter/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @palmares/redis-emitter
 
+## 1.0.2
+
+### Patch Changes
+
+- asdasdsa
+- Updated dependencies
+  - @palmares/events@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Testing if changesets and if custom script works
+- Updated dependencies
+  - @palmares/events@1.0.1
+
 ## 1.0.0
 
 ### Minor Changes