+++ libs/zod-schema/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @palmares/zod-schema
 
+## 0.1.2
+
+### Patch Changes
+
+- asdasdsa
+- Updated dependencies
+  - @palmares/serializers@1.0.2
+
+## 0.1.1
+
+### Patch Changes
+
+- Testing if changesets and if custom script works
+- Updated dependencies
+  - @palmares/serializers@1.0.1
+
 ## 0.1.0
 
 ### Minor Changes